Correct typo in random_graphs
Change the filehandler logic in the io.py routines so that
if path is a filehandle (actually anything with
a "write"/"readlines" attribute), path is used as
the filehandle.  Otherwise we try to open a file
with name stored in path as before.

--HG--
extra : convert_revision : svn%3A3ed01bd8-26fb-0310-9e4c-ca1a4053419f/networkx/trunk%40136