Fixes #37.
Passing a digraph to a connected component function now
raises an exception.  This applies the terminology decsion that the
term "connected" should be only be used for digraphs.

This solution may not be the final answer on implementation
but should prevent some confusion for now.

It might be useful to implement reachability and strong connectiviy
for digraphs.

--HG--
extra : convert_revision : svn%3A3ed01bd8-26fb-0310-9e4c-ca1a4053419f/networkx/trunk%40286