Fix normalized laplacian when the graph contains selfloops.
When G contains selfloops, the normalized laplacian is defined as NL = D^{-1/2}
(D - A) D^{-1/2}, with A the adjencency matrix and D a diagonal matrix with the
sum of the rows of A in the diagonal. The current NetworkX implementation use D
with the degree of the nodes in the diagonal, which count selfloops twice (this
is the right convension in general, but not the correct one for computing the
normalized laplacian).