Improve performance of descendants and ancestors
The `descendants` function for DAGs was slow because it performed a
check of whether the input graph was a DAG. This is now off by default.

In addition, the inverse function, `ancestors`, would previously
reverse the entire graph to find ancestors using BFS. A modification
to the `bfs_edges()` function allows reversing the BFS itself, avoiding
the cost of the full graph reversal.