Don't copy subgraphs in cuthill_mckee_ordering().
The connected_component_subgraphs() function used in the cuthill_mckee_ordering() makes a copy.  That causes new node objects to be created which might not be desirable when returning a node ordering if the hash changes when a new object is created. Fix by not copying nodes when forming a subgraph of connected nodes.