Don't import Graph Atlas generator if the interpreter is Jython.
The graph_atlas_g function is larger than the maximum method size
limit in Java (64K). This is just a workaround to get networkx to work
with Jython, and obviously it would be preferable to take all of the
Graph Atlas data out of the body of the function, and save it to a
separate data file.