Never import the atlas graph generator by default.
(the docstring for the generators.atlas module already said: "Because
of its size, this module is not imported by default.".)

Also, skip tests involving the graph atlas if running under
Jython. Many other tests still result in a failure or an error with
this interpreter, but at least networkx can be imported.