Inputs in bipartite generator funcs must be lists
The documentation states that the degree sequences provided as inputs to the
`bipartite_configuration_model` function may be lists or "iterators". However,
the code includes calls to `len()`, multiple iterations over the same iterable,
and access of elements at specific indices, all of which fail for generator
expressions.

Changes the documentation in the functions in this module to be clearer about
this requirement.