Improve interface to flow algorithms.
This commit introduces backwards incompatible changes to the interface
of flow algorithms:

1. Split maximum_flow into maximum_flow (returns a tuple flow_value, flow_dict)
and maximum_flow_value (returns only the value).

2. Split minimum_cut into minimum_cut (returns a tuple cut_value, partition)
and minimum_cut_value (returns only the value).

3. Remove the short versions max_flow and min_cut.

4. Remove flord_fulkerson_flow and ford_fulkerson_and_auxiliary

5. Remove the parameter `legacy` from ford_fulkerson. It now returns a
residual network as all other flow algorithms. I also tried to remove it
but doing so requires quite a few changes in mincost.py because the code,
tests, and the docstrings still relay on the its obsolete convention
for flow_dict.