Refactor connectivity package.
Use the new algorithms and the new interfaces provided by the flow package.
The code is quite clearer, and the new algorithms and interface provide
a significant speed up in running time. For some problems, this code is one order
of magnitude faster (see #1102 for bechmarks). So far the code is still
backwards compatible, but we should discuss some possible non backward
compatible improvements.