Improve auxiliary digraphs for connectivity functions.
1. The node mapping needed for node connectivity and minimum node cuts
is now a graph attribute of the auxiliary digraph. Thus there is no
need for a mapping parameter in the local version of these functions.

2. Change the parameter name for the auxuliary digraph from `aux_digraph`
to `auxiliary`. Also be consistent on the auxiliary digraph variable name
in the code. Now it is always `H`.

3. Added small sanity check for the auxiliary digraph for node connectivity.
If a digraph is passed as a paramater for reuse, check that it has a graph
attribute with the node mapping. If not we raise instead of rebuild the
auxiliary digraph.