Remove all_pairs_node_connectivity_matrix function.
With the addition of the example of how to compute local node connectivity
among all pairs of nodes reusing the data structures (added in the previous
commit in the docstrings of local_node_connectivity). We can remove this
function (which is the only one that has numpy dependency in the
connectivity package).