Do not allow weighted computations in minimum_st_edge_cut.
All connectivity and cut algorithms are supposed to work on unit
capacity networks. minimum_st_edge_cut was the only exception, because
it uses the new interface to flow algorithms, is potentially able to compute
weighted cutsets. However, this complicated the implementation and could
result in a NetworkXUnbounded exception if an user used as argument a
graph with some but not all edges with an attribute called capacity. Since
minimum_cut computes weighted cuts, there is no need to duplicate
functionality here.