Allows parallel edges when converting from numpy.
Forces from_numpy_matrix() and from_scipy_sparse_matrix() to create a
multigraph with parallel edges if the input adjacency matrix has integer
entries and the `create_using` keyword argument is an instance of
`MultiGraph`. In this case, each positive integer entry in the adjacency
matrix is interpreted as the number of parallel edges, each with weight
1, joining two vertices. Otherwise, if any of these conditions is not
met, the function behaves as it previously did: an integer entry is
interpreted as the weight of a single edge joining the two vertices.

This is not a backwards compatible change: it changes the behavior of
the function if the input is an integer matrix and `create_using` is a
multigraph.