Adds a "walk power" operation.
This commit adds the `walk_power` function in the module
`networkx.algorithms.operators.product`. This function returns the graph
in which there is an edge joining two nodes for each walk of a given
length connecting those nodes. There are two implementations, one with
SciPy and one without, with the former preferred.

It also creates the `networkx.algorithms.walks` module, containing
functions related to walks in a graph; the non-SciPy implementation of
the `walk_power` function requires functions for computing the number of
walks connecting two nodes, which may be of independent value.