Better proposed structure for OrderedGraph and co.
This supercedes #1268 as a proposed structure for OrderedGraph and
ThinGraph class structures. It is based on discussion from #980 and
has also been discussed in #1181, #1244, #1267, #1267

This implementation includes examples (in docs and in tests) for 1)
ordered nodes, 2) ordered nodes and edges and 3) thin (low RAM) graphs
that don't need edge attributes [not related to ordered nodes].

In this proposal, the dict-like arguments are assigned in a new class
structure rather than assigning them to an instance of the SpecialGraph
class. Something similar to:

class OrderedGraph(SpecialGraph):
    node_dict_factory=OrderedDict
G=OrderedGraph()

If all goes well with this version (including reasonable speed)
it might be reasonable to include it in the base classes Graph, etc.
I think @ysitu prefers a more flexible base class than adding yet
more graph classes, and all else equal (or almost equal) I agree.

This pull request may be useful for people who wish OrderedGraph
or ThinGraph and don't want to wait for 2.0