Allows explicit conversion with parallel edges
Adds the `parallel_edges` keyword argument to the `from_numpy_matrix`
and `from_scipy_sparse_matrix`. This allows the user to specify whether
to interpret an integer matrix as parallel edges in a multigraph or
single weighted edges in a multigraph.

This restores the default behavior of these functions that was changed
by pull request #1305.