Iterate over DataFrame.values, fix doc example
This improves performance substantially. DataFrame.iterrows incurs a large
overhead as it creates a new Series for each row. Modified this function to
loop over DataFrame.values instead, which is much more efficient. Each row is
now returned as as Numpy array, so we have to use index matching instead.

Fixed a couple of doc string examples as well that were causing problems in
the doc build.