Moves unit tests out of main network/ package
Before, unit tests were distributed within the `networkx` package, with
a distinct directory containing unit tests for each sub-package. Doing
this adds a great number of files that do not need to be distributed
with the NetworkX code.

This commit moves those test files to the `tests/` directory, which is
outside the `networkx` package, so they will no longer be distributed
(for example, in PyPI) with NetworkX.