Add all minimum-size k node cutsets algorithm.
This PR implements Kanevsky's algorithm for all minimum-size separating
node sets in an undirected graph as a generator of node sets.

The most tricky thing about the implementation is that for reusing the
residual network and the auxiliary graph for node connectivity (which
is key for speed) we have to keep removing and adding edges to them
instead of regenerating both each time we have to modify the input
graph (the algorithm prescrives to add an edge each time we find
a node cutset to avoid finding it again).

This PR uses almost as is a function from SAGE (see `compute_antichains`),
maybe we should ask permission to it's author to include it given that SAGE
is GPL and we are BSD?