Add k-(node)-components algorithm.
Add Moody and White algorithm for for identifying k-components in a graph,
which is based on Kanevsky's algorithm for finding all minimum-size node
cut-sets (implemented in `all_node_cuts` function #1391). This algorithm
consists in keep cutting a graph until we obtain either a complete or a
trivial graph.

`all_node_cuts` is doing most of the work here, the only tricky
implementation thing is the generation of partitions. The authors say that
``Nodes in a cutset belong to both sides of the induced cut.'', but an
induced cut can split a graph in more than two sides. Thus, after a cut,
we have to merge the node sets that (after adding the nodes in the cut set)
share at least k nodes.

I found out that this is very related to set consolidation, which is
implemented using connected components. This implementation is slower
than the iterative version from rosettacode.org because of the overhead
of building a graph and calling nx.connected_components, but we perform
the same number of intersection operations.

Also, it's possible that subsets of an already detected k-component appear
at a level k. Checking for this in the main while loop penalizes the common
case. Thus we also have to _consolidate all connectivity levels in
_reconstruct_k_components.