Add fast approximation for k-(node)-components.
This is an approximation approach to compute the k-component structure
of a graph (#1414 is the exact, and slow, version of this). I wrote a
paper to explore this approach in detail, see http://arxiv.org/abs/1503.04476

The logic of the approximation algorithm for computing the `k`-component
structure is based on repeatedly applying simple and fast algorithms
for `k`-cores and biconnected components in order to narrow down the
number of pairs of nodes over which we have to compute White and Newman's
approximation algorithm for finding node independent paths (implemented at #1405).
More formally, this algorithm is based on Whitney's theorem, which states
an inclusion relation among node connectivity, edge connectivity, and
minimum degree for any graph G. This theorem implies that every
`k`-component is nested inside a `k`-edge-component, which in turn,
is contained in a `k`-core. Thus, this algorithm computes node independent
paths among pairs of nodes in each biconnected part of each `k`-core,
and repeats this procedure for each `k` from 3 to the maximal core number
of a node in the input graph.

Because, in practice, many nodes of the core of level `k` inside a
bicomponent actually are part of a component of level k, the auxiliary
graph needed for the algorithm it's likely to be very dense. Thus, we use
a complement graph data structure to save memory (`_AntiGraph` see #599
for an earlier dicussion about this). AntiGraph only stores information
of the edges that are *not* present in the actual auxiliary graph.
However, when applying algorithms to this complement graph data structure,
it behaves as if it were the dense version. So it can be used directly
in several NetworkX algorithms, this version is only tested for the
algorithms needed here: `connected_components`, `k-cores`, and
`biconnected_components`. I used a `ThinGraph` for building the `AntiGraph`
and the memory consumption is quite small compared with a regular `Graph`.

I think `AntiGraph` is not useful enough to be a top level class in NetworkX,
but it could be a good example of a `Graph` subclass. I'll add to the
examples folder.

This PR supersedes and closes #580.