Adds function for computing a quotient graph.
The quotient graph of a graph `G` with respect to an equivalence
relation `R` on the nodes of `G` is the graph whose vertex set is the
equivalence classes `R` and where there is an edge joining class `c` to
class `d` if some vertex in `c` is adjacent to some vertex in `d`.

This commit places that function in a new module,
`networkx/algorithms/minors.py`, for functions related to graph minors.