Refactor and improve the documentation for connected_components.
Make connected_components a generator of sets of nodes, it was a generator
of list of nodes. See #1501 for discussion.

Use a plain BFS to compute them instead of using the single source
shortest path function. The latter also does a BFS but keeps track of the
distances. This is only slightly faster for small graphs (arround 5% in my
tests), but for larger graphs the speed up is more relevant (around 1.8x
faster). Also I think it's conceptually better. As @ysitu said, we have a
tendency to abuse shortest paths functions for reachability.

Also improved biconnected components examples, docstrings, and code.