Make strongly CC return a generator of sets, and improve docs and tests.
Following #1501 work, now strongly_connected_components and friends
yield sets of nodes. Added examples to the documentation for getting a
sorted list of components, and also the largest. Improved tests and
raised coverage to 100%.

Fixed a docstring example for `simple_cycles` that was relaying on
the output order of SCC.