Fixes approximate dominating set greedy choice.
Before, the greedy algorithm for approximate dominating set chose each
node based on the sum of the weights in the closed neighborhood of the
node. Now it just uses the weight of the node. This fixes an issue that
causes, for example, the return value for a star graph to be all nodes
other than the center node.