Fixes and extends Wiener index, vitality functions
Previously, the closeness vitality function was using an incorrect
definition of the Wiener index, counting each orientation of a pair of
nodes in an undirected graph twice. This commit corrects the behavior of
the Wiener index function and therefore the closeness vitality function.

Correcting the Wiener index function also corrects another case for the
closeness vitality function: when removing a node causes a graph to
become disconnected. By definition, the closeness vitality of a node
whose removal causes the graph to become disconnected should be negative
infinity. This commit fixes the Wiener index function to return positive
infinity on a disconnected graph.

Furthermore, this commit improves the closeness vitality function so
that it does not mutate the graph in order to compute the vitality
value.

Finally, this places the Wiener index function in its own module,
documents it, and makes it available to the end user.