Remove some tests for k_components approximation.
This algorithm is based on a node connectivity approximation that only
guarantees a lower bound for node connectivity. Even though some tests only
fail sporadically in Travis, and we cannot reproduce them, it's better to just
remove the problematic tests based on random graphs. Those tests were designed
for the exact implementation of the algorithm in `connectivity/kcomponents.py`.