# This is a combination of 3 commits. # The first commit's message is: # This is a combination of 2 commits. # The first commit's message is:
Remove some tests for k_components approximation.

This algorithm is based on a node connectivity approximation that only
guarantees a lower bound for node connectivity. Even though some tests only
fail sporadically in Travis, and we cannot reproduce them, it's better to just
remove the problematic tests based on random graphs. Those tests were designed
for the exact implementation of the algorithm in `connectivity/kcomponents.py`.

# The 2nd commit message will be skipped:

#	Added a test.

# The 2nd commit message will be skipped:

#	Fixes.

# The 3rd commit message will be skipped:

#	Import order.