Makes girvan_newman return communities as sets.
Previously, the community-finding algorithm `girvan_newman` returned a
list of tuples of lists of nodes. Now the function returns an iterator
over tuples of sets of nodes, each set representing a community. This
matches the behavior of the other community-finding functions.

This also prevents computing the connected components of the graph an
additional time when yielding the tuples of communities.