Fix type inference to be python3 compatible
Due to issues with long not existing in python 3, we need to reassign it
in the case of python 3 to int.

In addition we have to pull in lxml in order to parse the generated XML
to ensure it has the proper number of keys. We cannot compare to the
fixture due to mixed results in the ordering between python versions.