Moves triad_graph function to generators package.
Previously, the `triad_graphs` function in the
`networkx.algorithms.triads` module, returned a dictionary containing a
graph representing each possible triad. This commit moves that function
to a new module, `networkx.generators.triads`, and adds previously
missing test code for that function. It also modifies the function to
return a single graph instead of a dictionary of graphs. The dictionary
can now be generated by the code

    from networkx.algorithms.triads import TRIAD_NAMES
    triad_graphs = {name: triad_graph(name) for name in TRIAD_NAMES}