Reinstate v1.10 layout except center. Fix bugs
Center and scale are now available for all layout functions.
And points are correctly centered. The improvement is that
the default center value is sometimes scale/2 and sometimes 0.
This matches previous defaults and makes sense too -- for
circular layout the center is naturally the origin while for
random layout the domain is more naturally [0, scale).

Added tests for corner cases. Added fruchterman_reingold_layout
to the docs.