Makes balanced unary tree generator return path.
Previously, the `balanced_tree` generator function always returned a
path graph on two nodes instead of a path graph of the appropriate
length when provided a branching factor of one. This corrects that error
so the function now returns the path graph on `h + 1` nodes, where `h`
is the specified height of the tree, when the branching factor is one.