Allows copying a graph without attribute data.
Adds the `with_data` keyword argument to the `Graph.copy()` method,
which specifies whether the graph, node, and edge data will be included
in the copy of the graph object.

This commit also updates algorithms that include a call to
`Graph.copy()` to use `with_data=False` if they really do not need a
copy of the data.