Fixes tests for maximal matching.
Previously the tests for maximal matchings were testing that a given set
of edges was an edge cover, *not* a matching. In addition, they were not
testing maximality of the matching. This updates the tests to correctly
test that the `maximal_matching()` function returns a valid maximal
matching.

As part of this fix, this commit also introduces two public functions,
`is_matching()` and `is_maximal_matching()`, which decide whether a
given set of edges is a valid matching or maximal matching,
respectively.