Fix all_node_cuts corner cases: cycle and complete graphs.
This PR fixes #1875 . For cycle graphs and complete graphs all_node_cuts
was giving incorrect or incomplete results. The most tricky case is cycle
graphs because handling them as a part of the implementation of Kanevsky's
algorithm requires checks inside tight loops that would penalize the common
case.

Instead in this PR cycle and complete graphs are handled separately. It is
cheap to test if the input graph is any of both, and the solution of finding
all node cuts is a lot easier and faster than in the general case.

Also added tests for both cycle and complete graphs.