Fix minimum_st_edge_cut documentation.
A left over in the docstring from the last refactor made somewhat
not clear that the functions in cuts.py do not consider weights
and compute cuts in terms only of number of nodes or edges. Also
added a reference to `minimum_cut` for the users interested in
weighted edge cuts.