Fix DiGraph() docstring to reflect input flexibility
The existing documentation on the DiGraph() constructor function says that it can only take "an edge list or a NetworkX graph object" as well as a few third-party formats. This isn't true, since the data is just passed to the to_networkx_graph() function, which can actually accept a far wider variety of input formats - including dict of dicts and dict of lists.