Add flow based node and edge dijoint paths.
This is a flow based implementation of edge and node disjoint
paths. We compute the maximum flow between source and target
on an auxiliary directed network. This auxiliary network is
the same used in the flow based edge and node connectivity
algorithms. See `networkx.algorithms.connectivity.utils`
for details. The saturated edges in the residual network
after running the maximum flow algorithm correspond to edge
or node disjoint paths between source and target in the
original network.

The user can specify the maximum number of paths to yield,
but we also take advantatge of the fact the number of disjoint
paths is bounded by the degree of source and target to speed
up the computation, especially in sparse graphs.

This function handles both directed and undirected graphs,
and can use all flow algorithms from NetworkX flow package.

This PR supersedes #2053.