Moves Graph Atlas to data file.
This commit moves the edge lists of the graphs in the graph atlas from a
Python list defined in `networkx/generators/atlas.py` to a text file in
the same directory. This means that accessing the list of atlas graphs
will cause a file read.

This also adds a new function, `graph_atlas(i)`, that outputs only the
`i`th graph from the atlas.