Modified Graph.add_node and test_add_node.
Graph.add_node mutated the attr_dict parameter and assigned it directly
to the graph's node dictionary, creating potential for unexpected
behavior. Creating a new dictionary based on the supplied attr_dict (and
updated with the supplied kwargs) avoids this problem -- also had to
update the test for adding nodes, since sending an empty sequence no
longer raises an error (which seems reasonable).