Add Boykov Kolmogorov algorithm for maximum flow problems.
This algorithm has worse case complexity `O(n^2m|C|)` where |C| is the
cost of the minimum cut [1], but it is faster than other algorithms with
theoretically better worse case complexities for some problems, such as
vision and image processing, and maybe others.

For the case of connectivity problems in networks with higly skewed
degree distributions (eg social networks) this algorithm doesn't run
faster than Edmonds-Karp but after reading about it I thought that it
was really cool and decided to implement it.

This is still WIP as the marking heuristic described in [2] is
not yet implemented. According to the authors it speeds up the running
time considerably.

[1] Boykov, Y., & Kolmogorov, V. (2004). An experimental comparison of
min-cut/max-flow algorithms for energy minimization in vision.
http://www.csd.uwo.ca/~yuri/Papers/pami04.pdf
[2] Vladimir Kolmogorov. Graph-based Algorithms for Multi-camera
Reconstruction Problem. PhD thesis, Cornell University, CS Department,
2003. pp. 109-114.
https://pub.ist.ac.at/~vnk/papers/thesis.pdf