Implement marking heuristic for Boykov-Kolmogorov algorithm.
This heuristic uses an estimate of the distance to the root for
every node in both search trees, and a timestamp reflecting
when the distance was computed. As time goes by this estimate
gets outdated. In the adoption stage it is necessary to routinely
do checks in the search trees to see if a given node has source
or target as a root of its tree (see `_has_valid_root`), when
this checks are performed the distance estimate for each node
is updated. See [1] (pp. 109-114) for further details.

The aim of this heuristic is to shorten the paths that are
discovered in the growth stage, because after the first one,
there is no guarantee that other discovered paths will be the
short. In the adoption stage the search trees can be modified
in strange ways without applying this heuristic.

As a result the running time in many practical problems is
greatly improved.

[1] Vladimir Kolmogorov. Graph-based Algorithms for Multi-camera
Reconstruction Problem. PhD thesis, Cornell University, CS Department,
2003.
https://pub.ist.ac.at/~vnk/papers/thesis.pdf