Add search trees to a graph attribute and improve documentation.
A nice feature of the Boykov-Kolmogorov algorithm is that a partition
of the nodes that defines a minimum cut can be easily computed based
on the search trees used during the algorithm. These trees are now
stored in the graph attribute `trees` of the residual network.

Also added an example in the docstrings on how to compute the
partition that defines a minimum cut from the source and target
search trees, along with a refernece to the marking heuristic,
which is what makes this algorithm fast in practice for many
problems.