Adds beam search traversal algorithm with example
This commit adds the generalization of breadth-first search known as
"beam search", in which only a subset of "good" neighbors are enqueued
when visiting a new node. This refactors the breadth-first search
functions to call a generic BFS helper function that allows an arbitrary
successors function.

This also adds an example application, which is the progressive widening
beam search.