Fix fruchterman reingold bug and add more tests to layouts.
Incorrect generator expression syntax when evaluating the maximum of
the input pos dictionary. This bug only tiggered when a pos dict
was passed as argument to fruchterman_reingold_layout. This commit
adds tests for this case.

Also added more tests for layouts, which are currently quite sparse.
When doing so I also found and fixed a bug in circular_layout
corner case (graph with only one node).