Implemented proper arrows.
The main challenge in this task was to ensure a consistent scaling of
draw elements (nodes, node edges, edges, and arrow heads). Matplotlib
has 3 different scales / length units (figure, axis and data) and
different draw elements are tied to different scales. In the previous
implementation draw elements with different scales were
mixed (c.f. node_size=300 but linewidth=1.), making it impossible to
design a decent layout a priori, and in particular to choose sensible
values for arrow heads and arrow head offsets to prevent occlusion by/of
nodes. In this new implementation, all draw elements are derived from
matplotlib.patches, and are hence of uniform
scale (c.f. node_size=3. and linewidth=0.5). As far as I can tell, the
only compatibility breaking change was to scrap the
draw_networkx_edges() argument 'style', as there is no equivalent in the
matplotlib.patches class.