Refactor Dinitz' algorithm implementation.
Reimplemented the DFS iteratively instead of recursively. This
speeds up the running time considerably, specially for problems
in which the paths between source and sink are long, such as big
and sparse networks (eg social networks). For dense networks, in
which the paths are typically short, the running time is very
simmilar to the recursive implementation.

The speed up in the running time makes now possible to use Dinitz'
algorithm for testing the large graphs in `test_maxflow_large_graph.py`.
Thus we are now using all flow algorithms in all flow tests.