Fix test ordering in readwrite
Some interesting finds here and one that I don't understand.

Using OrderedGraph is helpful in writing tests of output to get deterministic order. But OrderedDict __repr__ is 'OrderedDict()' so I made an ugly modification in generate_edgelist.

I don't understand why the sparse6 write requires the sorted order to pass the tests.  There might be a bug here.

I fixed a bug in graph6 where the nodes needed to be sorted