Makes write_graph6 less memory-intensive.
This commit changes the implementation of the `write_graph6` function so
that it makes better use of iterators and writes directly to file
instead of buffering the entire encoding of the graph as a string in
memory before writing. This commit also removes the `generate_graph6`
function, since it's functionality can be simulated by using the
`write_graph6` function to write to a file.

This change requires updating the code to explicitly require `bytes`
when reading and writing graph6 data. This means updating `sparse6` code
as well.

Finally, this commit standardizes the graph6 and sparse6
encoding/decoding functions into four functions each: `to_graph6_bytes`,
`from_graph6_bytes`, `read_graph6`, and `write_graph6`, and the
corresponding functions for sparse6.