Fixes for k-components tests and node connectivity approximation doctests (#5)
* Use octahedral instead of icosahedral for connectivity approx docs.

The approximation algorithm for node connectivity fails on the
icosahedral graph depending on node iteration order. We use the
octahedral graph instead as example in the docstrings.

* Fix k-components tests avoiding hard-coded node labels.

In order to support the new dict implementation of python 3.6 we
have to avoid hard-coding node labels for the k-components of the
example graphs in k-components tests. We now test the number of
k-components found, their number of nodes, and their node
connectivity to make sure that the algorithms are correct.