Use faster random geometric graph implementation.
Replace the existing random geometric graph generator with one
that uses a k-d tree to efficiently determine which nodes are
adjacent, so the algorithm is now `O(n)` instead of `O(n^2)`.

Fall back to non-scipy `O(n^2)` version when scipy is not
available.

Use p-norm number as parameter instead of function. The
scipy KDTree requires a number instead of distance function.