Raise an Exception for disconnected Graphs in bipartite.sets
This pull request fixes #2127

When a disconnected graph is passed to `bipartite.sets` more than one
coloring solution is possible. This led to surprising results and hard
to diagnose bugs (see #2127) in algorithms that used `bipartite.sets`,
such as bipartite matching and friends. These algorithms now have a new
optional parameter that allows the user to specify the nodes in one of
the two bipartite sets, and thus allows to use disconnected graphs in
these algorithms without ambiguity.

I've added a new exception `AmbiguousSolution` in order to be more
explicit when we raise it. It might have more use cases beyond
`bipartite.sets`.

The changes in this pull request are backwards compatible for connected
graphs as the new parameter is optional. For disconnected graphs is not
backward compatible, but since the assignation of nodes to a bipartite
set depends on iteration order, that code was already subtly broken
as it would yield different results on different python versions.